move to src for cli
