+++ cli/src/database/get.go
